refactor: capture socket and move to capture dir
